TamkeenAI_CareerSystem/requirements.txt
flask==3.1.0
flask-cors==4.0.0
gunicorn==21.2.0
pymongo==4.12.0
pyjwt==2.10.1
flask-jwt-extended==4.7.1
python-dotenv==1.0.1
numpy==1.26.4
pandas==2.2.1
scikit-learn==1.4.2
requests==2.32.3
huggingface_hub==0.29.3
openai==1.72.0
nltk==3.8.1
